yunyou/yunyou_backend/requirements.txt
# Django and DRF
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
mysqlclient>=2.2.0

# Auth
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# AI Integration
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.1.0
redis>=5.0.0

# Backup
django-dbbackup>=4.0.0

# API Documentation
drf-spectacular>=0.27.0
